fix:medication filter ui

+++ app/src/main/java/com/example/careplus/data/filter_model/FilterMedications.kt
@@ -1,14 +1,14 @@
 package com.example.careplus.data.filter_model
 
 data class FilterMedications(
-    val active: Boolean? = null ,
-    val caregiver_id: Int? = null,
-    val diagnosis_id: Int? = null,
-    val doctor_id: Int? = null,
+    val active: Boolean? = null,
+    val caregiver_id: Long? = null,
+    val diagnosis_id: Long? = null,
+    val doctor_id: Long? = null,
     val end_date: String? = null,
-    val form_id: Int? = null,
-    val patient_id: Int? = null,
-    val route_id: Int? = null,
+    val form_id: Long? = null,
+    val patient_id: Long? = null,
+    val route_id: Long? = null,
     val search: String? = null,
     val start_date: String? = null,
     //for pagination
